integration test, fix user_id res 0, adjust field to gmtoffset, get user before delete, gmtoffset float, update readme, transform path param type

+++ migrations/20240220184257_ceate_user_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS public.users (
     first_name varchar NOT NULL,
     last_name varchar NOT NULL,
     birthday date NOT NULL,
-	gmt_offset smallint NOT NULL,
+	gmt_offset float NOT NULL,
 	last_run timestamp,
 	last_send timestamp,
 	CONSTRAINT users_pk PRIMARY KEY (id)
